refactor(wallet-sdk): split sdk.ts into file-per-domain api factories (Phase 3.5)
Checkpoint feedback: sdk.ts was ~300 lines and grows by a domain each
phase. Each domain now owns {domain}/{domain}-api.ts holding the *Api type
and a create*Api(deps) factory; sdk.ts shrinks to config + the composition
root (one create*Api call per domain).

createAccountsApi returns {api, repository, cache} so the root can wire
cross-domain deps (the user domain's WriteUserRepository and its upsert
accounts write-back) against the same instances without reaching through
the transitional `internal` escape hatch — the root must not depend on a
surface documented as removable.

createLazyEncryption moves to encryption.ts where its collaborators live;
package-root export is preserved via the existing barrel. AccountsApi and
UserApi stay exported from sdk.ts as type re-exports.
