+++ docs/wallet-sdk-extraction.md
@@ -116,7 +116,8 @@ db-singleton → accounts-core → sdk-root → user-types → user-core → use
 | 3.1 | `sdk/phase3-user-types` · c2eda5b6 | `User`/`FullUser`/`GuestUser`/`UserProfile`/`UpdateUser` + 3 predicates → SDK `user/user.ts`; `AccountService`'s structural `UserDefaultAccounts` replaced with the real `User`; web `user.ts` → shim |
 | 3.2 | `sdk/phase3-user-core` · 4312041e | `ReadUserRepository`/`WriteUserRepository` (`upsert_user_with_accounts` RPC)/`ReadUserDefaultAccountRepository` + `UserService` + `UserCache`/`userQueryOptions`/`createUserChangeHandlers` → SDK (verbatim, import remaps only); web repo/service files → shims keeping hooks one more chunk |
 | 3.3 | `sdk/phase3-user-surface` · f4ab912f | curated `sdk.user` surface: `queryOptions/getCached/upsert/update/setDefaultAccount` (cache write-backs absorbed; `upsert` records accounts via new `AccountsCache.set`); `internal = {readRepository, writeRepository, service, cache, changeHandlers}`; `ensureUserData` + accept-terms + verify-email + receive-token migrated; `useReadUserRepository`/`useWriteUserRepository`/`useUserService` deleted. `ReadUserDefaultAccountRepository` deliberately NOT in the root: its only consumer is the server-side lnurl path (per-request server db + `LNURL_SERVER_SPARK_MNEMONIC`); a root instance would wrongly bind the logged-in user's mnemonic |
-| 3.4 | `sdk/phase3-user-surface` · (review fixup) | checkpoint feedback: `getCachedOrThrow` removed from `sdk.user` — whether a missing cached user is exceptional is caller policy, and accounts exposes no orThrow either. SDK keeps the `getCached(): User \| null` primitive; web's `getUserFromCacheOrThrow` (user-hooks) wraps it for protected-layout contexts where missing user = bug |
+| 3.4 | `sdk/phase3-user-surface` · ca4937df | checkpoint feedback: `getCachedOrThrow` removed from `sdk.user` — whether a missing cached user is exceptional is caller policy, and accounts exposes no orThrow either. SDK keeps the `getCached(): User \| null` primitive; web's `getUserFromCacheOrThrow` (user-hooks) wraps it for protected-layout contexts where missing user = bug |
+| 3.5 | `sdk/phase3-user-surface` · (file-per-domain split) | checkpoint feedback: `sdk.ts` split — each domain owns `{domain}/{domain}-api.ts` with the `*Api` type + `create*Api(deps)` factory; `sdk.ts` is just config + the composition root. `createAccountsApi` returns `{api, repository, cache}` so the root wires cross-domain deps (user's WriteUserRepository, upsert accounts write-back) WITHOUT reaching through `internal`; `createLazyEncryption` moved to `encryption.ts`. Pattern for every future domain: new file + one `create*Api` call in the root |
 
 ## Remaining roadmap
